+++ vitest.shared.ts
@@ -13,6 +13,15 @@ export default {
   resolve: {
     alias: {
       'react-querybuilder': path.resolve(root, 'packages/react-querybuilder/src'),
+      '@react-querybuilder/core/parseJSONata': path.resolve(
+        root,
+        'packages/core/src/utils/parseJSONata'
+      ),
+      '@react-querybuilder/core/parseJsonLogic': path.resolve(
+        root,
+        'packages/core/src/utils/parseJsonLogic'
+      ),
+      '@react-querybuilder/core/parseSQL': path.resolve(root, 'packages/core/src/utils/parseSQL'),
       '@react-querybuilder/core': path.resolve(root, 'packages/core/src'),
       '@rqb-testing': path.resolve(root, 'utils/testing'),
       '@rqb-dbpool': path.resolve(root, 'utils/testing/pglite'),